// src/types/app.ts
export type OtherBaseURLKey = 'demo';

export interface ServiceConfigItem {
  /** The backend service base url */
  baseURL: string
  /** The proxy pattern of the backend service base url */
  proxyPattern: string
}

export interface OtherServiceConfigItem extends ServiceConfigItem {
  key: OtherBaseURLKey
}

/** The backend service config */
export interface ServiceConfig extends ServiceConfigItem {
  /** Other backend service config */
  other: OtherServiceConfigItem[]
}

export interface SimpleServiceConfig extends Pick<ServiceConfigItem, 'baseURL'> {
  other: Record<OtherBaseURLKey, string>
}